// src/pages/Employee/types.ts
export interface Employee {
  id?: number;
  fullName: string;
  dateOfBirth: string;
  gender: string;
  nationalId: string;
  phone: string;
  email: string;
  address: string;
  jobTitle: string;
  dateOfHire: string;
  education: string;
  bankAccount: string;
  salary: number;
  appraisalRecords?: string;
  disciplinaryActions?: string;
}